change p0nki's code

+++ src/main/java/tk/valoeghese/fc0/client/ClientPlayer.java
@@ -224,7 +224,8 @@ public class ClientPlayer {
 				return new RaycastResult(new TilePos(pos.x,pos.y,pos.z),face);
 			}
 		}
-		return new RaycastResult(new TilePos((int) end.x, (int) end.y, (int) end.z),null);
+
+		return new RaycastResult(new TilePos(end.x, end.y, end.z),null);
 	}
 
 	public void tick() {
